Some fixes to map generation and camera.
I had to add in the line: rooms = new List<Room>(); in order to get the
code to run for me. After that I then tweaked the camera tracking to
work, as well as made is so the player and enemies wouldn't rotate when
they collided with walls and stuff (as rotation would be odd with later
parts of the game being grid based, this can easily change once
animations/models are in). There is still a semi-random array overflow
error in the room generation.