Player attack phase added
The player now has an attack phase after its move phase which can be
ended either by attacking an enemy or by hitting enter again. This may
change as it is a little clunky. The player movement seems to not fully
line up with the grid, making some enemies spawn in a "no mans land"
where they cannot be attacked.