player now has their model attached
I attached the player model to the player. I implemented this as a
separate object that is a child of Player called PlayerRenderer, which
has mesh and meshRenderer components. This method allows us to make any
kind of visual adjustments without worrying about altering behavior
(e.g. the player's model can be scaled and rotated without affecting the
player object's collider). I also tweaked the camera and lighting a bit
so things look nice and 3d