PlayerUnitManager script... manages player units
There is now behavior in the PlayerManager script. It holds a list of
player units. When a PlayerUnit is instantiated, it registers itself to
the PlayerUnitmanager by adding itself to the list. PlayerUnitManager
waits for a unit to register, then sets activeUnit to that unit and
starts activeUnit's PlayerTurn coroutine. It should be pretty simple at
this point to press a button and toggle through units by stopping active
unit's PlayerTurn, changing activeUnit to the next element of the unit
list, then starting that unit's PlayerTurn.