It's and initiative system, but only kinda.
An initiative system is set up, however it uses every enemy in the game,
not just per room. Which floods the turn counter, and makes it basically
function just like it did before. This should be fixed once only enemies
in the player's combat room are added to the TurnManager's combatant
list.