Fixed sound from clipping on camera movement
When the camera moved a large distance, the music would experience a
bit of weird fracturing. This was due to the 3d panning effect built
into fmod (so we were hearing a note or two as if the source of the
sound was distant and in a different direction). Removed the 3d panner
to fix this